// src/app/core/allworker_api/model/education.ts
/**
 * API Candidats
 * API pour la gestion des candidats
 *
 * OpenAPI spec version: v1
 * Contact: contact@votreapi.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface Education { 
    readonly id?: number;
    candidat: number;
    institution: string;
    degree: string;
    startDate?: string;
    endDate?: string;
}